Fixed AmPieChart, AmSlicedChart and AmChart - getJso , jso, WrapperUtils.unwrap not the same in JSNI as seen in getGroupPercent code in AmSlicedChart. - getJso in AmPieChart cast to AmPieCharJSO and used for setGroupedPulled and getGroupedPulled demonstrates non requirement of JSNI code in AmPieChart for setter and getters.

+++ src/com/amcharts/api/IsAmSlicedChart.java
@@ -60,9 +60,9 @@ public interface IsAmSlicedChart extends IsAmChart
 
 	public abstract void setGroupedTitle( String groupedTitle );
 
-	public abstract Float getGroupPercent();
+	public abstract Double getGroupPercent();
 
-	public abstract void setGroupPercent( Float groupPercent );
+	public abstract void setGroupPercent( Double groupPercent );
 
 	public abstract Float getHideLabelsPercent();
